Version:3.0 - Cloud Run Basic Dockerized/requirements.txt
gunicorn==22.0.0
Flask==3.0.3
requests==2.32.3
google-generativeai==0.7.2
vertexai==1.49.0
markdown==3.6
pelican[markdown]==4.9.1
python-dotenv==1.0.1
openai==1.37.1